refactor: move EnrollSection component to shared folder
