+++ src/modules/HomePage/components/organisms/index.tsx
@@ -1,6 +1,6 @@
 export { default as Alumni } from './AlumniSection';
 export { default as AvailableProgramsSection } from './AvailableProgramsSection';
-export { default as EnrollSection } from './EnrollSection';
 export { default as HeroSection } from './HeroSection';
 export { default as JoiningProcessFlow } from './JoiningProcessFlow';
 export { default as ValueSection } from './ValueSection';
+export { EnrollSection } from '@/shared/components/molecules';
